refactor(loon-core): split path read internals
