+++ crates/loon-core/src/path/mod.rs
@@ -3,4 +3,5 @@ pub(crate) mod intent;
 pub(crate) mod mutation;
 pub(crate) mod planner;
 pub(crate) mod query;
+mod read;
 pub(crate) mod tombstone;
